[front] refactor: extract conversation auto-scroll hook
